requirements.txt
fastapi
docker
gitpython
ngrok
uvicorn
tinydb